refactor: keep server implementation independent of signup rpc
